move htmlToFormattedText/htmlToSegments into @frogpond/html-lib
Agent-Logs-Url: https://github.com/StoDevX/AAO-React-Native/sessions/0dfb8beb-ac67-4080-abc7-2672c66fa82f
